+++ tests/testthat/test-istr.R
@@ -41,7 +41,6 @@ test_that("if `companies` lacks crucial columns, errors gracefully", {
   expect_error(istr(bad, scenarios, inputs), crucial)
 })
 
-# TODO: Common test in both istr and pstr because both use same scenario dataset
 test_that("if `scenarios` lacks crucial columns, errors gracefully", {
   companies <- istr_companies |> slice(1)
   scenarios <- istr_scenarios
@@ -92,15 +91,17 @@ test_that("if `inputs` lacks crucial columns, errors gracefully", {
 
 test_that("thresholds yield expected low, medium, and high risk categories", {
   companies <- tibble(
-    company_id = "cta-commodity-trading-austria-gmbh_00000005215384-001",
+    company_id = "a",
+    tilt_sector = "any",
     clustered = "any",
-    activity_uuid_product_uuid = "any",
-    ei_activity_name = "any",
-    unit = "unit"
+    activity_uuid_product_uuid = "any"
   )
 
   inputs <- tibble(
     activity_uuid_product_uuid = "any",
+    input_activity_uuid_product_uuid = "any",
+    input_tilt_sector = "any",
+    input_tilt_subsector = "any",
     type = "ipr",
     sector = "total",
     subsector = "energy",
@@ -149,13 +150,12 @@ test_that("thresholds yield expected low, medium, and high risk categories", {
   expect_equal(0, filter(out, risk_category == "high")$value)
 })
 
-# TODO: This test should be changed to handle `NA` values in the `value` column
 test_that("outputs values in proportion", {
-  companies <- istr_companies |> slice(3)
+  companies <- istr_companies |> slice(1)
   scenarios <- istr_scenarios
   inputs <- istr_inputs
   out <- istr(companies, scenarios, inputs)
-  expect_true(all(out$value <= 1.0))
+  expect_true(all(na.omit(out$value) <= 1.0))
 })
 
 test_that("each company has risk categories low, medium, and high (#215)", {
@@ -177,10 +177,9 @@ test_that("grouped_by includes the type of scenario", {
   expect_true(all(grepl(.type, unique(out$grouped_by))))
 })
 
-# TODO: This test should be changed to handle `NA` values in the `value` column
 test_that("with type ipr, for each company and grouped_by value sums 1 (#216)", {
   .type <- "ipr"
-  companies <- istr_companies |> slice(3)
+  companies <- istr_companies |> slice(1)
   scenarios <- istr_scenarios |> filter(type == .type)
   inputs <- istr_inputs |> filter(type == .type)
 
@@ -188,7 +187,7 @@ test_that("with type ipr, for each company and grouped_by value sums 1 (#216)", 
   sum <- out |>
     summarize(value_sum = sum(value), .by = c("companies_id", "grouped_by"))
 
-  expect_true(all(sum$value_sum == 1))
+  expect_true(all(na.omit(sum$value_sum) == 1))
 })
 
 test_that("values sum 1 or are NA if a company does or doesn't match (#176)", {
@@ -203,14 +202,16 @@ test_that("values sum 1 or are NA if a company does or doesn't match (#176)", {
 
   companies <- tibble(
     company_id = c("a", "b"),
+    tilt_sector = c("any", "any"),
     clustered = "x",
-    activity_uuid_product_uuid = c("x", "y"),
-    ei_activity_name = "x",
-    unit = "x"
+    activity_uuid_product_uuid = c("x", "y")
   )
 
   inputs <- tibble(
     activity_uuid_product_uuid = c("x", "y"),
+    input_activity_uuid_product_uuid = c("any", "any"),
+    input_tilt_sector = c("any", "any"),
+    input_tilt_subsector = c("any", "any"),
     type = c("x", "y"),
     sector = c("x", "y"),
     subsector = c("x", "y"),
@@ -233,14 +234,16 @@ test_that("values sum 1 or are NA if a company does or doesn't match (#176)", {
 test_that("no matches yield the expected prototype (#176)", {
   companies <- tibble(
     company_id = c("a", "b"),
+    tilt_sector = c("any", "any"),
     clustered = "x",
-    activity_uuid_product_uuid = c("x", "y"),
-    ei_activity_name = "x",
-    unit = "x"
+    activity_uuid_product_uuid = c("x", "y")
   )
 
   inputs <- tibble(
     activity_uuid_product_uuid = c("x", "y"),
+    input_activity_uuid_product_uuid = c("any", "any"),
+    input_tilt_sector = c("any", "any"),
+    input_tilt_subsector = c("any", "any"),
     type = c("x", "y"),
     sector = c("x", "y"),
     subsector = c("x", "y"),
@@ -265,10 +268,9 @@ test_that("no matches yield the expected prototype (#176)", {
   expect_equal(unique(unmatched$value), NA_real_)
 })
 
-# TODO: This test should be changed to handle `NA` values in the `value` column
 test_that("with type weo, for each company and grouped_by value sums 1 (#308)", {
   .type <- "weo"
-  companies <- istr_companies |> slice(3)
+  companies <- istr_companies |> slice(1)
   scenarios <- istr_scenarios |> filter(type == .type)
   inputs <- istr_inputs |> filter(type == .type)
 
@@ -276,20 +278,22 @@ test_that("with type weo, for each company and grouped_by value sums 1 (#308)", 
   sum <- out |>
     summarize(value_sum = sum(value), .by = c("companies_id", "grouped_by"))
 
-  expect_true(all(sum$value_sum == 1))
+  expect_true(all(na.omit(sum$value_sum) == 1))
 })
 
 test_that("error if a `type` has all `NA` in `sector` & `subsector` (#310)", {
   companies <- tibble(
     company_id = "a",
+    tilt_sector = "any",
     clustered = "x",
-    activity_uuid_product_uuid = "f",
-    ei_activity_name = "x",
-    unit = "x"
+    activity_uuid_product_uuid = "f"
   )
 
   inputs <- tibble(
     activity_uuid_product_uuid = "f",
+    input_activity_uuid_product_uuid = "any",
+    input_tilt_sector = "any",
+    input_tilt_subsector = "any",
     type = "b",
     sector = "c",
     subsector = "d",
@@ -334,29 +338,27 @@ test_that("the thresholds are in the range 0 to 1", {
 })
 
 test_that("NA in the reductions column yields `NA` in risk_category at product level", {
-  companies <- tibble(
-    company_id = "1",
-    type = "a",
-    sector = "b",
-    subsector = "c",
-    clustered = "any",
-    activity_uuid_product_uuid = "x",
-    tilt_sector = "x",
-    tilt_subsector = "x",
-  )
   scenarios <- tibble(
+    reductions = NA,
     scenario = "2",
     sector = "b",
     subsector = "c",
     year = 2020,
-    reductions = NA,
     type = "a",
+  )
+
+  companies <- tibble(
+    company_id = "1",
+    clustered = "any",
+    activity_uuid_product_uuid = "x",
+    tilt_sector = "x",
   )
 
   inputs <- tibble(
     activity_uuid_product_uuid = "x",
-    input_activity_uuid_product_uuid = "y",
-    input_reference_product_name = "y",
+    input_activity_uuid_product_uuid = "any",
+    input_tilt_sector = "any",
+    input_tilt_subsector = "any",
     input_unit = "y",
     input_isic_4digit = "y",
     type = "a",
@@ -371,13 +373,11 @@ test_that("NA in the reductions column yields `NA` in risk_category at product l
 test_that("NA in the reductions column should be ignored from the value calculations", {
   companies <- tibble(
     company_id = "1",
-    type = "a",
-    sector = "b",
-    subsector = "c",
+    ei_activity_name = "x",
+    unit = "x",
     clustered = "any",
     activity_uuid_product_uuid = "x",
     tilt_sector = "x",
-    tilt_subsector = "x",
   )
   scenarios <- tibble(
     scenario = "2",
@@ -390,8 +390,9 @@ test_that("NA in the reductions column should be ignored from the value calculat
 
   inputs <- tibble(
     activity_uuid_product_uuid = "x",
-    input_activity_uuid_product_uuid = "y",
-    input_reference_product_name = "y",
+    input_activity_uuid_product_uuid = "any",
+    input_tilt_sector = "any",
+    input_tilt_subsector = "any",
     input_unit = "y",
     input_isic_4digit = "y",
     type = "a",